Move things to tests sources, lint
