+++ massa-bootstrap/src/tests/tools.rs
@@ -1,6 +1,7 @@
 // Copyright (c) 2022 MASSA LABS <info@massa.net>
 
-use crate::settings::{BootstrapConfig, IpType};
+use crate::settings::{BootstrapClientConfig, BootstrapConfig, IpType};
+use crate::BootstrapServerMessage;
 use bitvec::vec::BitVec;
 use massa_async_pool::AsyncPoolChanges;
 use massa_async_pool::{test_exports::get_random_message, AsyncPool};
@@ -10,14 +11,14 @@ use massa_consensus_exports::{
     },
     export_active_block::{ExportActiveBlock, ExportActiveBlockSerializer},
 };
-use massa_db_exports::{DBBatch, ShareableMassaDBController};
+use massa_db_exports::{DBBatch, ShareableMassaDBController, StreamBatch};
 use massa_executed_ops::{
     ExecutedDenunciations, ExecutedDenunciationsChanges, ExecutedDenunciationsConfig, ExecutedOps,
     ExecutedOpsConfig,
 };
 use massa_final_state::test_exports::create_final_state;
 use massa_final_state::{FinalState, FinalStateConfig};
-use massa_hash::Hash;
+use massa_hash::{Hash, HASH_SIZE_BYTES};
 use massa_ledger_exports::{LedgerChanges, LedgerEntry, SetOrKeep, SetUpdateOrDelete};
 use massa_ledger_worker::test_exports::create_final_ledger;
 use massa_models::block::BlockDeserializerArgs;
@@ -25,18 +26,20 @@ use massa_models::bytecode::Bytecode;
 use massa_models::config::{
     BOOTSTRAP_RANDOMNESS_SIZE_BYTES, CONSENSUS_BOOTSTRAP_PART_SIZE, ENDORSEMENT_COUNT,
     MAX_ADVERTISE_LENGTH, MAX_ASYNC_MESSAGE_DATA, MAX_ASYNC_POOL_LENGTH,
-    MAX_BOOTSTRAP_ASYNC_POOL_CHANGES, MAX_BOOTSTRAP_BLOCKS,
-    MAX_BOOTSTRAP_ERROR_LENGTH, MAX_CONSENSUS_BLOCKS_IDS, MAX_DATASTORE_ENTRY_COUNT,
-    MAX_DATASTORE_KEY_LENGTH, MAX_DATASTORE_VALUE_LENGTH, MAX_DEFERRED_CREDITS_LENGTH,
-    MAX_DENUNCIATIONS_PER_BLOCK_HEADER, MAX_DENUNCIATION_CHANGES_LENGTH,
-    MAX_EXECUTED_OPS_CHANGES_LENGTH, MAX_EXECUTED_OPS_LENGTH, MAX_FUNCTION_NAME_LENGTH,
-    MAX_LEDGER_CHANGES_COUNT, MAX_OPERATIONS_PER_BLOCK, MAX_OPERATION_DATASTORE_ENTRY_COUNT,
-    MAX_OPERATION_DATASTORE_KEY_LENGTH, MAX_OPERATION_DATASTORE_VALUE_LENGTH, MAX_PARAMETERS_SIZE,
-    MAX_PRODUCTION_STATS_LENGTH, MAX_ROLLS_COUNT_LENGTH, MIP_STORE_STATS_BLOCK_CONSIDERED,
-    PERIODS_PER_CYCLE, THREAD_COUNT, MAX_BOOTSTRAPPED_NEW_ELEMENTS_SIZE,
+    MAX_BOOTSTRAPPED_NEW_ELEMENTS_SIZE, MAX_BOOTSTRAP_ASYNC_POOL_CHANGES, MAX_BOOTSTRAP_BLOCKS,
+    MAX_BOOTSTRAP_ERROR_LENGTH, MAX_BOOTSTRAP_FINAL_STATE_PARTS_SIZE, MAX_CONSENSUS_BLOCKS_IDS,
+    MAX_DATASTORE_ENTRY_COUNT, MAX_DATASTORE_KEY_LENGTH, MAX_DATASTORE_VALUE_LENGTH,
+    MAX_DEFERRED_CREDITS_LENGTH, MAX_DENUNCIATIONS_PER_BLOCK_HEADER,
+    MAX_DENUNCIATION_CHANGES_LENGTH, MAX_EXECUTED_OPS_CHANGES_LENGTH, MAX_EXECUTED_OPS_LENGTH,
+    MAX_FUNCTION_NAME_LENGTH, MAX_LEDGER_CHANGES_COUNT, MAX_OPERATIONS_PER_BLOCK,
+    MAX_OPERATION_DATASTORE_ENTRY_COUNT, MAX_OPERATION_DATASTORE_KEY_LENGTH,
+    MAX_OPERATION_DATASTORE_VALUE_LENGTH, MAX_PARAMETERS_SIZE, MAX_PRODUCTION_STATS_LENGTH,
+    MAX_ROLLS_COUNT_LENGTH, MIP_STORE_STATS_BLOCK_CONSIDERED, PERIODS_PER_CYCLE, THREAD_COUNT,
 };
 use massa_models::denunciation::DenunciationIndex;
 use massa_models::node::NodeId;
+use massa_models::prehash::PreHashSet;
+use massa_models::version::Version;
 use massa_models::{
     address::Address,
     amount::Amount,
@@ -59,7 +62,8 @@ use massa_signature::KeyPair;
 use massa_time::MassaTime;
 use massa_versioning::versioning::{MipStatsConfig, MipStore};
 use num::rational::Ratio;
-use rand::Rng;
+use rand::rngs::SmallRng;
+use rand::{Rng, SeedableRng};
 use std::collections::{HashMap, HashSet};
 use std::str::FromStr;
 use std::{
@@ -563,4 +567,358 @@ pub fn get_peers(keypair: &KeyPair) -> BootstrapPeers {
             listeners2,
         ),
     ])
+}
+
+fn gen_export_active_blocks<R: Rng>(rng: &mut R) -> ExportActiveBlock {
+    let keypair = KeyPair::generate(0).unwrap();
+    let block = gen_random_block(&keypair, rng)
+        .new_verifiable(BlockSerializer::new(), &keypair)
+        .unwrap();
+    let parents = (0..32)
+        .map(|_| (gen_random_block_id(rng), rng.gen()))
+        .collect();
+    ExportActiveBlock {
+        block,
+        parents,
+        is_final: rng.gen_bool(0.5),
+    }
+}
+
+fn gen_new_stream_batch<R>(rng: &mut R) -> StreamBatch<Slot>
+where
+    R: Rng,
+{
+    let change_id = gen_random_slot(rng);
+    let batch_size = rng.gen_range(0..MAX_BOOTSTRAP_FINAL_STATE_PARTS_SIZE);
+    let new_elements_size = rng.gen_range(0..batch_size);
+    let mut new_elements = BTreeMap::new();
+    let mut size = 0;
+    loop {
+        let key = gen_random_vector(MAX_DATASTORE_KEY_LENGTH as usize, rng);
+        let val = gen_random_vector(MAX_DATASTORE_VALUE_LENGTH as usize, rng);
+        if size + (key.len() as u64) + (val.len() as u64) > new_elements_size {
+            break;
+        }
+        size += key.len() as u64;
+        size += val.len() as u64;
+        new_elements.insert(key, val);
+    }
+    let mut updates_on_previous_elements = BTreeMap::new();
+    loop {
+        let key = gen_random_vector(MAX_DATASTORE_KEY_LENGTH as usize, rng);
+        let mut s = key.len();
+        let val = if rng.gen_bool(0.5) {
+            let v = gen_random_vector(MAX_DATASTORE_VALUE_LENGTH as usize, rng);
+            s += v.len();
+            Some(v)
+        } else {
+            None
+        };
+        if (size + (s as u64)) > batch_size {
+            break;
+        }
+        size += s as u64;
+        updates_on_previous_elements.insert(key, val);
+    }
+    StreamBatch {
+        new_elements,
+        updates_on_previous_elements,
+        change_id,
+    }
+}
+
+fn gen_random_string<R: Rng>(max: usize, rng: &mut R) -> String {
+    let res = gen_random_vector(max, rng);
+    res.into_iter()
+        .map(|c| char::from_u32(((c as u32) % (122 - 65)) + 65).unwrap())
+        .collect()
+}
+
+fn gen_random_vector<R: Rng>(max: usize, rng: &mut R) -> Vec<u8> {
+    let nb = rng.gen_range(0..=max);
+    let mut res = vec![0; nb];
+    rng.fill_bytes(&mut res);
+    res
+}
+
+fn stream_batch_equal<T: PartialOrd + Ord + PartialEq + Eq + Clone + std::fmt::Debug>(
+    s1: &StreamBatch<T>,
+    s2: &StreamBatch<T>,
+) -> bool {
+    let mut new_elements_equal = true;
+    for (key1, val1) in s1.new_elements.iter() {
+        if let Some(val2) = s2.new_elements.get(key1) {
+            new_elements_equal &= val1 == val2;
+        } else {
+            new_elements_equal = false;
+            break;
+        }
+    }
+    let mut update_prevels_equal = true;
+    for (key1, val1) in s1.updates_on_previous_elements.iter() {
+        if let Some(val2) = s2.updates_on_previous_elements.get(key1) {
+            update_prevels_equal &= val1 == val2;
+        } else {
+            update_prevels_equal = false;
+            break;
+        }
+    }
+    new_elements_equal && update_prevels_equal && (s1.change_id == s2.change_id)
+}
+
+fn gen_random_hash<R: Rng>(rng: &mut R) -> Hash {
+    let bytes: [u8; HASH_SIZE_BYTES] = rng.gen();
+    Hash::from_bytes(&bytes)
+}
+
+fn gen_random_slot<R: Rng>(rng: &mut R) -> Slot {
+    Slot {
+        period: rng.gen(),
+        thread: rng.gen_range(0..32),
+    }
+}
+
+fn gen_random_block_id<R: Rng>(rng: &mut R) -> BlockId {
+    BlockId::generate_from_hash(gen_random_hash(rng))
+}
+
+fn gen_random_block<R: Rng>(keypair: &KeyPair, rng: &mut R) -> Block {
+    let slot = gen_random_slot(rng);
+    let parents: Vec<BlockId> = (0..32).map(|_| gen_random_block_id(rng)).collect();
+    let mut endorsements = vec![];
+    for index in 0..rng.gen_range(1..ENDORSEMENT_COUNT) {
+        let endorsement = Endorsement {
+            index,
+            slot,
+            endorsed_block: parents[slot.thread as usize],
+        };
+
+        let endorsement = endorsement
+            .new_verifiable(EndorsementSerializer::new(), &keypair)
+            .unwrap();
+        // assert!(endorsement.verify_signature().is_ok(), "Endorsement signature invalid");
+        // assert!(endorsement.check_invariants().is_ok(), "Endorsememnt invariants are invalid");
+        endorsements.push(endorsement);
+    }
+
+    let mut denunciations = vec![];
+    for index in 0..rng.gen_range(0..(MAX_DENUNCIATIONS_PER_BLOCK_HEADER as usize)) {
+        // TODO    Denunciations generation
+    }
+
+    let header = BlockHeader {
+        current_version: rng.gen(),
+        announced_version: rng.gen(),
+        slot,
+        parents,
+        operation_merkle_root: gen_random_hash(rng),
+        endorsements,
+        denunciations,
+    }
+    .new_verifiable(BlockHeaderSerializer::new(), &keypair)
+    .unwrap();
+    let mut operations = vec![];
+    for _ in 0..rng.gen_range(0..MAX_OPERATIONS_PER_BLOCK) {
+        let op = OperationId::new(gen_random_hash(rng));
+        operations.push(op);
+    }
+    Block { header, operations }
+}
+
+impl BootstrapServerMessage {
+    pub fn generate<R: Rng>(rng: &mut R) -> Self {
+        let variant = rng.gen_range(0..6);
+        match variant {
+            0 => {
+                let t: u64 = rng.gen();
+                // Taken from INSTANCE_LEN in version.rs
+                let vi: String = (0..4)
+                    .map(|_| char::from_u32(rng.gen_range(65..91)).unwrap())
+                    .collect();
+                let major: u32 = rng.gen();
+                let minor: u32 = rng.gen();
+                let version =
+                    Version::from_str(format!("{}.{}.{}", vi, major, minor).as_str()).unwrap();
+                let server_time = MassaTime::from_millis(t);
+                BootstrapServerMessage::BootstrapTime {
+                    server_time,
+                    version,
+                }
+            }
+            1 => {
+                let peer_nb = rng.gen_range(0..100);
+                let mut peers_list = vec![];
+                for _ in 0..peer_nb {
+                    peers_list.push((
+                        PeerId::from_public_key(KeyPair::generate(0).unwrap().get_public_key()),
+                        HashMap::new(),
+                    ));
+                }
+                let peers = BootstrapPeers(peers_list);
+                BootstrapServerMessage::BootstrapPeers { peers }
+            }
+            2 => {
+                let state_part = gen_new_stream_batch(rng);
+                let versioning_part = gen_new_stream_batch(rng);
+                let mut final_blocks = vec![];
+                let block_nb = rng.gen_range(5..100); //MAX_BOOTSTRAP_BLOCKS);
+                for _ in 0..block_nb {
+                    final_blocks.push(gen_export_active_blocks(rng));
+                }
+                let nb = rng.gen_range(0..100);
+                let mut consensus_outdated_ids = PreHashSet::default();
+                for _ in 0..nb {
+                    consensus_outdated_ids.insert(gen_random_block_id(rng));
+                }
+                let consensus_part = BootstrapableGraph { final_blocks };
+                let last_start_period = rng.gen();
+                let last_slot_before_downtime = if rng.gen_bool(0.5) {
+                    Some(Some(gen_random_slot(rng)))
+                } else {
+                    None
+                };
+                let slot = gen_random_slot(rng);
+                BootstrapServerMessage::BootstrapPart {
+                    slot,
+                    state_part,
+                    versioning_part,
+                    consensus_part,
+                    consensus_outdated_ids,
+                    last_start_period,
+                    last_slot_before_downtime,
+                }
+            }
+            3 => BootstrapServerMessage::BootstrapFinished,
+            4 => BootstrapServerMessage::SlotTooOld,
+            5 => BootstrapServerMessage::BootstrapError {
+                error: gen_random_string(MAX_BOOTSTRAP_ERROR_LENGTH as usize, rng),
+            },
+            _ => unreachable!(),
+        }
+    }
+
+    pub fn equals(&self, other: &BootstrapServerMessage) -> bool {
+        match (self, other) {
+            (
+                BootstrapServerMessage::BootstrapTime {
+                    server_time: t1,
+                    version: v1,
+                },
+                BootstrapServerMessage::BootstrapTime {
+                    server_time: t2,
+                    version: v2,
+                },
+            ) => (t1 == t2) && (v1 == v2),
+            (
+                BootstrapServerMessage::BootstrapPeers { peers: p1 },
+                BootstrapServerMessage::BootstrapPeers { peers: p2 },
+            ) => p1 == p2,
+            (
+                BootstrapServerMessage::BootstrapPart {
+                    slot: s1,
+                    state_part: state1,
+                    versioning_part: v1,
+                    consensus_part: c1,
+                    consensus_outdated_ids: co1,
+                    last_start_period: lp1,
+                    last_slot_before_downtime: ls1,
+                },
+                BootstrapServerMessage::BootstrapPart {
+                    slot: s2,
+                    state_part: state2,
+                    versioning_part: v2,
+                    consensus_part: c2,
+                    consensus_outdated_ids: co2,
+                    last_start_period: lp2,
+                    last_slot_before_downtime: ls2,
+                },
+            ) => {
+                let state_equal = stream_batch_equal(state1, state2);
+                let versionning_equal = stream_batch_equal(v1, v2);
+                let mut consensus_equal = true;
+                if c1.final_blocks.len() != c2.final_blocks.len() {
+                    return false;
+                }
+                for (n, active_block1) in c1.final_blocks.iter().enumerate() {
+                    let active_block2 = c2.final_blocks.get(n).unwrap();
+                    consensus_equal &= active_block1.parents == active_block2.parents;
+                    consensus_equal &= active_block1.is_final == active_block2.is_final;
+                    consensus_equal &=
+                        active_block1.block.serialized_data == active_block2.block.serialized_data;
+                }
+                (s1 == s2)
+                    && state_equal
+                    && versionning_equal
+                    && consensus_equal
+                    && (co1 == co2)
+                    && (lp1 == lp2)
+                    && (ls1 == ls2)
+            }
+            (
+                BootstrapServerMessage::BootstrapFinished,
+                BootstrapServerMessage::BootstrapFinished,
+            ) => true,
+            (BootstrapServerMessage::SlotTooOld, BootstrapServerMessage::SlotTooOld) => true,
+            (
+                BootstrapServerMessage::BootstrapError { error: e1 },
+                BootstrapServerMessage::BootstrapError { error: e2 },
+            ) => e1 == e2,
+            _ => false,
+        }
+    }
+}
+
+pub fn bootstrap_client_parametric_test<F>(niter: usize, regressions: Vec<u64>, test_fct: F)
+where
+    F: Fn(&BootstrapClientConfig, &mut SmallRng),
+{
+    use massa_models::config::*;
+    let config = BootstrapClientConfig {
+        rate_limit: std::u64::MAX,
+        max_listeners_per_peer: MAX_LISTENERS_PER_PEER as u32,
+        endorsement_count: ENDORSEMENT_COUNT,
+        max_advertise_length: MAX_ADVERTISE_LENGTH,
+        max_bootstrap_blocks_length: MAX_BOOTSTRAP_BLOCKS,
+        max_operations_per_block: MAX_OPERATIONS_PER_BLOCK,
+        thread_count: THREAD_COUNT,
+        randomness_size_bytes: BOOTSTRAP_RANDOMNESS_SIZE_BYTES,
+        max_bootstrap_error_length: MAX_BOOTSTRAP_ERROR_LENGTH,
+        max_new_elements_size: MAX_BOOTSTRAP_FINAL_STATE_PARTS_SIZE,
+        max_datastore_entry_count: MAX_DATASTORE_ENTRY_COUNT,
+        max_datastore_key_length: MAX_DATASTORE_KEY_LENGTH,
+        max_datastore_value_length: MAX_DATASTORE_VALUE_LENGTH,
+        max_async_pool_changes: MAX_BOOTSTRAP_ASYNC_POOL_CHANGES,
+        max_async_pool_length: MAX_ASYNC_POOL_LENGTH,
+        max_async_message_data: MAX_ASYNC_MESSAGE_DATA,
+        max_ledger_changes_count: MAX_LEDGER_CHANGES_COUNT,
+        max_changes_slot_count: 1000,
+        max_rolls_length: MAX_ROLLS_COUNT_LENGTH,
+        max_production_stats_length: MAX_PRODUCTION_STATS_LENGTH,
+        max_credits_length: MAX_DEFERRED_CREDITS_LENGTH,
+        max_executed_ops_length: MAX_EXECUTED_OPS_LENGTH,
+        max_ops_changes_length: MAX_EXECUTED_OPS_CHANGES_LENGTH,
+        mip_store_stats_block_considered: MIP_STORE_STATS_BLOCK_CONSIDERED,
+        max_denunciations_per_block_header: MAX_DENUNCIATIONS_PER_BLOCK_HEADER,
+        max_denunciation_changes_length: MAX_DENUNCIATION_CHANGES_LENGTH,
+    };
+    parametric_test(niter, config, regressions, test_fct)
+}
+
+pub fn parametric_test<F, T>(niter: usize, config: T, regressions: Vec<u64>, test_fct: F)
+where
+    F: Fn(&T, &mut SmallRng),
+{
+    for reg in regressions {
+        println!("[*] Regression {reg}");
+        let mut rng = SmallRng::seed_from_u64(reg);
+        test_fct(&config, &mut rng);
+    }
+
+    let mut seeder = SmallRng::from_entropy();
+    for n in 0..niter {
+        let new_seed: u64 = seeder.gen();
+        println!("[{n}] Seed {new_seed}");
+        let mut rng = SmallRng::seed_from_u64(new_seed);
+        test_fct(&config, &mut rng);
+    }
 }
